Moved some tuning functions from .hpp to .cpp
